framework: Fix fetch-url-list and fetch-urlall-list
The fetch-url-list and fetch-urlall-list targets are meant to produce
a list of URLs from which a port fetches its distfiles and patches.

Currently, those targets were not working as expected as they print
parts of the output meant for other targets like the fetch-list target.

For example:

    mateusz.piotrowski@server /ports/x11-servers/xwayland-devel$ make fetch-urlall-list
    [...]
    mkdir -p "xorg" &&
    -o xorg/proto-xorgproto-824001c947cb1962209c6a8f2c63c2637877220d_GL0.tar.gz
    [...]

This patch prevents do-fetch.sh from printing the "mkdir" line and the
"-o" line.

While here:

- Remove the outdated comment about escaping. There is no excessive
  escaping happening there anymore.
- Indent cases in the case statement to match the style of the rest of
  the file.

Sponsored by:	Klara Inc.