math/py-numpy: Introduce options FORTRAN and NOBLAS
NumPy documentation[1] states that it is fully supported to use NumPy's
internal fallback implementation of the BLAS and LAPACK libraries. Let's
add that option to the port.

The internal fallback routines are documented to be rather slow in
comparison to the typical library implementations. To address a
potential misconfiguration mistake of an unsuspecting user, note the
slowness of the NOBLAS implementation in the option description.

[1]: https://numpy.org/devdocs/building/blas_lapack.html

PR:		268348
Co-authored-by:	Mateusz Piotrowski <0mp@FreeBSD.org>