samba.mk: Add samba:self, SAMBA_LDB_MODULESDIR and SAMBA_IDMAP_MODULESDIR
The new argument, samba:self, is similar to samba:env. It allows a port
to reference samba.mk variables without adding dependencies on the Samba
port to *_DEPENDS. The difference is that samba:self is intended to be
used by Samba ports. This way we can be sure that Samba and its
consumers agree on the location of shared files and directories.

New *DIR variables are SAMBA_LDB_MODULESDIR and SAMBA_IDMAP_MODULESDIR
as required by security/sssd2.

Sponsored by:	Klara, Inc.